Added closure-library to get_deps.py

+++ get_deps.py
@@ -3,7 +3,12 @@
 import urllib
 import subprocess
 import os
+import argparse
 
+
+parser = argparse.ArgumentParser()
+parser.add_argument("-c", "--closure", help="Also install the closure-library dependency",action="store_true")
+args = parser.parse_args()
 #
 # Create dependency output directories
 #
@@ -26,10 +31,13 @@ if not os.path.exists("output/blockly"):
 # instructions step by step.  So cheating a little and pulling the binary here from their docs page.
 # Apache 2.0 License
 #
-print("Downloading opencv.js...")
-urllib.urlretrieve("https://docs.opencv.org/3.4.1/opencv.js","output/opencv/opencv.js")
-urllib.urlretrieve("https://raw.githubusercontent.com/opencv/opencv/3.4.1/README.md","output/opencv/README.md")
-urllib.urlretrieve("https://raw.githubusercontent.com/opencv/opencv/3.4.1/LICENSE","output/opencv/LICENSE")
+if os.path.exists("output/opencv/opencv.js"):
+    print("opencv.js already exists. Skipping downloading opencv...")
+else:
+    print("Downloading opencv.js...")
+    urllib.urlretrieve("https://docs.opencv.org/3.4.1/opencv.js","output/opencv/opencv.js")
+    urllib.urlretrieve("https://raw.githubusercontent.com/opencv/opencv/3.4.1/README.md","output/opencv/README.md")
+    urllib.urlretrieve("https://raw.githubusercontent.com/opencv/opencv/3.4.1/LICENSE","output/opencv/LICENSE")
 
 #
 # FileSaver.js (Currently using 1.3.4)
@@ -38,15 +46,28 @@ urllib.urlretrieve("https://raw.githubusercontent.com/opencv/opencv/3.4.1/LICENS
 # Using the minimized version to save space, but should be swappable.
 # MIT License
 #
-print("Downloading FileSaver.js...")
-urllib.urlretrieve("https://raw.githubusercontent.com/eligrey/FileSaver.js/1.3.4/LICENSE.md", "output/filesaver/LICENSE.md");
-urllib.urlretrieve("https://raw.githubusercontent.com/eligrey/FileSaver.js/1.3.4/README.md", "output/filesaver/README.md");
-urllib.urlretrieve("https://raw.githubusercontent.com/eligrey/FileSaver.js/1.3.4/FileSaver.min.js", "output/filesaver/FileSaver.js");
+if os.path.exists("output/filesaver/FileSaver.js"):
+    print("FileSaver.js already exists. Skipping downloading FileSaver.js...")
+else:
+    print("Downloading FileSaver.js...")
+    urllib.urlretrieve("https://raw.githubusercontent.com/eligrey/FileSaver.js/1.3.4/LICENSE.md", "output/filesaver/LICENSE.md");
+    urllib.urlretrieve("https://raw.githubusercontent.com/eligrey/FileSaver.js/1.3.4/README.md", "output/filesaver/README.md");
+    urllib.urlretrieve("https://raw.githubusercontent.com/eligrey/FileSaver.js/1.3.4/FileSaver.min.js", "output/filesaver/FileSaver.js");
 
 #
 # Blockly
 # Main: https://developers.google.com/blockly/
 # Source: https://github.com/google/blockly
 # For now I'm pulling all of Blockly down.
-print("Cloning Blockly...")
-subprocess.call(['git.exe', 'clone', 'https://github.com/google/blockly.git', 'output/blockly'])
+if os.path.exists("output/blockly/blockly_compressed.js"):
+    print("Blockly already exists. Skipping cloning Blockly...")
+else:
+    print("Cloning Blockly...")
+    subprocess.call(['git.exe', 'clone', 'https://github.com/google/blockly.git', 'output/blockly'])
+
+if args.closure:
+    if (os.path.exists("output/closure-library")):
+        print("closure-library already exists. Skipping cloning closure-library...")
+    else:
+        print("Cloning closure-library...")
+        subprocess.call(['git.exe', 'clone', 'https://github.com/google/closure-library.git', 'output/closure-library'])
